+++ controllers/capi_controller.go
@@ -29,8 +29,11 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
+	eks "sigs.k8s.io/cluster-api-provider-aws/controlplane/eks/api/v1beta1"
 	capa "sigs.k8s.io/cluster-api-provider-aws/v2/api/v1beta1"
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/cluster-api/util/patch"
+	"sigs.k8s.io/cluster-api/util/predicates"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
@@ -41,40 +44,64 @@ import (
 	"github.com/giantswarm/irsa-operator/pkg/key"
 )
 
-// CAPAClusterReconciler reconciles a CAPA AWSCluster object
-type CAPAClusterReconciler struct {
+// CAPIClusterReconciler reconciles a CAPA AWSCluster object
+type CAPIClusterReconciler struct {
 	client.Client
-	Log    logr.Logger
-	Scheme *runtime.Scheme
+	Log              logr.Logger
+	Scheme           *runtime.Scheme
+	WatchFilterValue string
 
 	Installation string
 	recorder     record.EventRecorder
 }
 
-// +kubebuilder:rbac:groups=infrastructure.cluster.x-k8s.io,resources=awscluster,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=infrastructure.cluster.x-k8s.io,resources=awscluster/status,verbs=get;update;patch
-// +kubebuilder:rbac:groups=infrastructure.cluster.x-k8s.io,resources=awscluster/finalizers,verbs=update
+// +kubebuilder:rbac:groups=infrastructure.cluster.x-k8s.io,resources=cluster,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=infrastructure.cluster.x-k8s.io,resources=cluster/status,verbs=get;update;patch
+// +kubebuilder:rbac:groups=infrastructure.cluster.x-k8s.io,resources=cluster/finalizers,verbs=update
+// +kubebuilder:rbac:groups=infrastructure.cluster.x-k8s.io,resources=awscluster,verbs=get;list;watch
+// +kubebuilder:rbac:groups=controlplane.cluster.x-k8s.io,resources=awsmanagedcontrolplane,verbs=get;list;watch
 // +kubebuilder:rbac:groups=core,resources=configmaps,verbs=get;patch
 
 // Reconcile is part of the main kubernetes reconciliation loop which aims to
 // move the current state of the cluster closer to the desired state.
 // For more details, check Reconcile and its Result here:
 // - https://pkg.go.dev/sigs.k8s.io/controller-runtime@v0.8.3/pkg/reconcile
-func (r *CAPAClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+func (r *CAPIClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	var err error
 	logger := r.Log.WithValues("namespace", req.Namespace, "cluster", req.Name)
 
-	logger.Info("Reconciling AWSCluster")
+	logger.Info("Reconciling CAPI Cluster")
 
-	cluster := &capa.AWSCluster{}
+	cluster := &capi.Cluster{}
 	if err := r.Get(ctx, req.NamespacedName, cluster); err != nil {
 		return ctrl.Result{}, microerror.Mask(client.IgnoreNotFound(err))
 	}
 
+	var roleIdentityName string
+	var region string
+
+	if isCAPA(cluster) {
+		awsCluster := &capa.AWSCluster{}
+		if err := r.Get(ctx, req.NamespacedName, awsCluster); err != nil {
+			return ctrl.Result{}, microerror.Mask(client.IgnoreNotFound(err))
+		}
+		roleIdentityName = awsCluster.Spec.IdentityRef.Name
+		region = awsCluster.Spec.Region
+
+	} else if isEKS(cluster) {
+		eksCluster := &eks.AWSManagedControlPlane{}
+		if err := r.Get(ctx, req.NamespacedName, eksCluster); err != nil {
+			return ctrl.Result{}, microerror.Mask(client.IgnoreNotFound(err))
+		}
+		roleIdentityName = eksCluster.Spec.IdentityRef.Name
+		region = eksCluster.Spec.Region
+	}
+
 	awsClusterRoleIdentity := &capa.AWSClusterRoleIdentity{}
-	err = r.Get(ctx, types.NamespacedName{Name: cluster.Spec.IdentityRef.Name}, awsClusterRoleIdentity)
+
+	err = r.Get(ctx, types.NamespacedName{Name: roleIdentityName}, awsClusterRoleIdentity)
 	if err != nil {
-		return ctrl.Result{}, microerror.Mask(fmt.Errorf("failed to get AWSClusterRoleIdentity object %q: %w", cluster.Spec.IdentityRef.Name, err))
+		return ctrl.Result{}, microerror.Mask(fmt.Errorf("failed to get AWSClusterRoleIdentity object %q: %w", roleIdentityName, err))
 	}
 
 	arn := awsClusterRoleIdentity.Spec.RoleArn
@@ -97,7 +124,7 @@ func (r *CAPAClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request)
 		ClusterNamespace: cluster.Namespace,
 		ConfigName:       key.ConfigName(cluster.Name),
 		Installation:     r.Installation,
-		Region:           cluster.Spec.Region,
+		Region:           region,
 		// This is a hack to allow CAPI clusters to drop the 'release.giantswarm.io/version' label.
 		ReleaseVersion: "20.0.0-alpha1",
 		SecretName:     key.SecretName(cluster.Name),
@@ -131,10 +158,10 @@ func (r *CAPAClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request)
 		controllerutil.RemoveFinalizer(cluster, key.FinalizerName)
 		err = patchHelper.Patch(ctx, cluster)
 		if err != nil {
-			logger.Error(err, "failed to remove finalizer from AWSCluster")
+			logger.Error(err, "failed to remove finalizer from Cluster")
 			return ctrl.Result{}, err
 		}
-		logger.Info("successfully removed finalizer from AWSCluster")
+		logger.Info("successfully removed finalizer from Cluster")
 
 		r.sendEvent(cluster, v1.EventTypeNormal, "IRSA", "IRSA bootstrap deleted")
 
@@ -151,10 +178,10 @@ func (r *CAPAClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request)
 			controllerutil.AddFinalizer(cluster, key.FinalizerName)
 			err = patchHelper.Patch(ctx, cluster)
 			if err != nil {
-				logger.Error(err, "failed to add finalizer on AWSCluster")
+				logger.Error(err, "failed to add finalizer on Cluster")
 				return ctrl.Result{}, microerror.Mask(err)
 			}
-			logger.Info("successfully added finalizer to AWSCluster")
+			logger.Info("successfully added finalizer to Cluster")
 		}
 
 		err := irsaService.Reconcile(ctx)
@@ -175,9 +202,10 @@ func (r *CAPAClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request)
 }
 
 // SetupWithManager sets up the controller with the Manager.
-func (r *CAPAClusterReconciler) SetupWithManager(mgr ctrl.Manager) error {
+func (r *CAPIClusterReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	err := ctrl.NewControllerManagedBy(mgr).
-		For(&capa.AWSCluster{}).
+		For(&capi.Cluster{}).
+		WithEventFilter(predicates.ResourceHasFilterLabel(r.Log, r.WatchFilterValue)).
 		Complete(r)
 	if err != nil {
 		return errors.Wrap(err, "failed setting up with a controller manager")
@@ -187,6 +215,14 @@ func (r *CAPAClusterReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return nil
 }
 
-func (r *CAPAClusterReconciler) sendEvent(cluster *capa.AWSCluster, eventtype, reason, message string) {
+func (r *CAPIClusterReconciler) sendEvent(cluster *capi.Cluster, eventtype, reason, message string) {
 	r.recorder.Eventf(cluster, v1.EventTypeNormal, reason, message)
+}
+
+func isCAPA(capiCluster *capi.Cluster) bool {
+	return capiCluster.Spec.InfrastructureRef != nil && capiCluster.Spec.InfrastructureRef.Kind == "AWSCluster"
+}
+
+func isEKS(capiCluster *capi.Cluster) bool {
+	return capiCluster.Spec.InfrastructureRef != nil && capiCluster.Spec.InfrastructureRef.Kind == "AWSManagedCluster"
 }